refactor(core): extract script types to script.rs
Move ScriptPubkey to dedicated script module.
