+++ src/core/mod.rs
@@ -1,5 +1,6 @@
 pub mod block;
 pub mod block_tree_entry;
+pub mod script;
 pub mod transaction;
 
 pub use block::{
@@ -7,7 +8,9 @@ pub use block::{
     TransactionSpentOutputs, TransactionSpentOutputsRef,
 };
 pub use block_tree_entry::BlockTreeEntry;
+pub use script::{ScriptPubkey, ScriptPubkeyRef};
 pub use transaction::{Transaction, TransactionRef, TxOut, TxOutRef};
 
 pub use block::{BlockSpentOutputsExt, CoinExt, TransactionSpentOutputsExt};
+pub use script::ScriptPubkeyExt;
 pub use transaction::{TransactionExt, TxOutExt};
